Test update (#140)
* Test Updates

Added Distance tests
Added Itinerary tests

* modified tests

* update distance test values

* Update to tests
#111, closes #128

* Update Itinerary.js

The bounds break the code, so I'm just gonna comment it out so you can still work on it

+++ server/src/main/java/com/tripco/t23/TIP/TIPDistance.java
@@ -22,13 +22,13 @@ import java.util.Map;
 public class TIPDistance extends TIPHeader {
   private Map origin;
   private Map destination;
-  private Float earthRadius;
+  private double earthRadius;
   private long distance;
 
   private final transient Logger log = LoggerFactory.getLogger(TIPDistance.class);
 
 
-  TIPDistance(int version, Map origin, Map destination, float earthRadius) {
+  TIPDistance(int version, Map origin, Map destination, double earthRadius) {
     this();
     this.requestVersion = version;
     this.origin = origin;
